move db logic to repo for consistent approach
